+++ src/serial_interface/serial_config.h
@@ -32,6 +32,6 @@ int uart_writes(struct uart_device_struct *dev, char *string);
 int uart_reads(struct uart_device_struct *dev, char *buf, size_t buf_len);
 int uart_reads_chunk(struct uart_device_struct *dev, char *buf, size_t buf_len);
 void uart_stop(struct uart_device_struct *dev);
-int uart_gps_write(struct uart_device_struct *, const uint8_t *, int);
+int uart_gps_write(struct uart_device_struct *, const uint8_t *, uint8_t);
 
 #endif
